+++ test/testpgpr.c
@@ -234,7 +234,7 @@ keyinfo(int argc, char **argv)
     if (subkey)
 	key = select_subkey(pubkey, pubkeyl, key, subkey);
     printf("Version: %d\n", pgprItemVersion(key));
-    printf("CreationTime: %d\n", pgprItemCreationTime(key));
+    printf("CreationTime: %lld\n", (long long)pgprItemCreationTime(key));
     printf("Algorithm: %d\n", pgprItemPubkeyAlgo(key));
     printf("AlgorithmInfo: %d\n", pgprItemPubkeyAlgoInfo(key));
     printf("UserID: %s\n", nullify(pgprItemUserID(key)));
@@ -346,7 +346,7 @@ siginfo(int argc, char **argv)
     }
     free(lints);
     printf("Version: %d\n", pgprItemVersion(sig));
-    printf("CreationTime: %d\n", pgprItemCreationTime(sig));
+    printf("CreationTime: %lld\n", (long long)pgprItemCreationTime(sig));
     printf("Algorithm: %d\n", pgprItemPubkeyAlgo(sig));
     printf("Hash: %d\n", pgprItemHashAlgo(sig));
     keyfp = pgprItemKeyFingerprint(sig, &keyfp_len, NULL);
